+++ src/rpc_client/json/mod.rs
@@ -7,6 +7,8 @@ use std::fmt::{Debug, Display, Formatter, LowerHex, UpperHex};
 
 pub mod requests;
 pub mod responses;
+#[cfg(test)]
+mod tests;
 
 macro_rules! bytes_array {
     ($name: ident, $size: expr) => {
@@ -74,6 +76,7 @@ macro_rules! bytes_array {
 bytes_array!(FixedSizeData, 32);
 bytes_array!(Hash, 32);
 bytes_array!(LogsBloom, 256);
+bytes_array!(StorageKey, 32);
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(transparent)]