+++ tests/tests.rs
@@ -1,5 +1,6 @@
 mod mock;
 
+use crate::mock::MockJsonRequestBody;
 use assert_matches::assert_matches;
 use candid::{CandidType, Decode, Encode, Nat, Principal};
 use evm_rpc::logs::{Log, LogEntry};
@@ -24,6 +25,7 @@ use pocket_ic::common::rest::{
 };
 use pocket_ic::{CanisterSettings, PocketIc, WasmResult};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use serde_json::json;
 use std::sync::Arc;
 use std::{marker::PhantomData, str::FromStr, time::Duration};
 
@@ -278,6 +280,15 @@ impl EvmRpcSetup {
         )
     }
 
+    pub fn eth_call(
+        &self,
+        source: RpcServices,
+        config: Option<evm_rpc_types::RpcConfig>,
+        args: evm_rpc_types::CallArgs,
+    ) -> CallFlow<MultiRpcResult<evm_rpc_types::Hex>> {
+        self.call_update("eth_call", Encode!(&source, &config, &args).unwrap())
+    }
+
     pub fn update_api_keys(&self, api_keys: &[(ProviderId, Option<String>)]) {
         self.call_update("updateApiKeys", Encode!(&api_keys).unwrap())
             .wait()
@@ -463,7 +474,7 @@ fn mock_request_should_succeed_with_request_headers() {
 
 #[test]
 fn mock_request_should_succeed_with_request_body() {
-    mock_request(|builder| builder.with_request_body(MOCK_REQUEST_PAYLOAD))
+    mock_request(|builder| builder.with_raw_request_body(MOCK_REQUEST_PAYLOAD))
 }
 
 #[test]
@@ -481,7 +492,7 @@ fn mock_request_should_succeed_with_all() {
                 (CONTENT_TYPE_HEADER_LOWERCASE, CONTENT_TYPE_VALUE),
                 ("Custom", "Value"),
             ])
-            .with_request_body(MOCK_REQUEST_PAYLOAD)
+            .with_raw_request_body(MOCK_REQUEST_PAYLOAD)
     })
 }
 
@@ -506,7 +517,9 @@ fn mock_request_should_fail_with_request_headers() {
 #[test]
 #[should_panic(expected = "assertion `left == right` failed")]
 fn mock_request_should_fail_with_request_body() {
-    mock_request(|builder| builder.with_request_body(r#"{"different":"body"}"#))
+    mock_request(|builder| {
+        builder.with_raw_request_body(r#"{"id":1,"jsonrpc":"2.0","method":"unknown_method"}"#)
+    })
 }
 
 #[test]
@@ -888,6 +901,61 @@ fn eth_send_raw_transaction_should_succeed() {
                 Hex32::from_str(MOCK_TRANSACTION_HASH).unwrap()
             ))
         );
+    }
+}
+
+#[test]
+fn eth_call_should_succeed() {
+    const ADDRESS: &str = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48";
+    const INPUT_DATA: &str =
+        "0x70a08231000000000000000000000000b25eA1D493B49a1DeD42aC5B1208cC618f9A9B80";
+
+    let setup = EvmRpcSetup::new().mock_api_keys();
+    for call_args in [
+        evm_rpc_types::CallArgs {
+            transaction: evm_rpc_types::TransactionRequest {
+                to: Some(ADDRESS.parse().unwrap()),
+                input: Some(INPUT_DATA.parse().unwrap()),
+                ..evm_rpc_types::TransactionRequest::default()
+            },
+            block: Some(evm_rpc_types::BlockTag::Latest),
+        },
+        evm_rpc_types::CallArgs {
+            transaction: evm_rpc_types::TransactionRequest {
+                to: Some(ADDRESS.parse().unwrap()),
+                input: Some(INPUT_DATA.parse().unwrap()),
+                ..evm_rpc_types::TransactionRequest::default()
+            },
+            block: None, //should be same as specifying Latest
+        },
+    ] {
+        for source in RPC_SERVICES {
+            let response = setup
+                .eth_call(source.clone(), None, call_args.clone())
+                .mock_http(MockOutcallBuilder::new(
+                    200,
+                    r#"{"jsonrpc":"2.0","result":"0x0000000000000000000000000000000000000000000000000000013c3ee36e89","id":1}"#,
+                )
+                    .with_request_body(
+                        MockJsonRequestBody::builder("eth_call")
+                            .with_params(json!(
+                            [
+                                {
+                                    "to": ADDRESS.to_lowercase(),
+                                    "input": INPUT_DATA.to_lowercase(),
+                                },
+                                "latest"
+                            ]
+                ))))
+                .wait()
+                .expect_consistent()
+                .unwrap();
+            assert_eq!(
+                response,
+                Hex::from_str("0x0000000000000000000000000000000000000000000000000000013c3ee36e89")
+                    .unwrap()
+            );
+        }
     }
 }